Move InsightlyServiceProvider to Insightly namespace
